Latest working script

+++ setup.sh
@@ -1,5 +1,43 @@
 #!/bin/sh
 
+generateDockItem() {
+	printf "%s%s%s%s%s%s%s%s%s%s%s%s" \
+		"<dict>" \
+			"<key>tile-data</key>" \
+			"<dict>" \
+				"<key>file-data</key>" \
+				"<dict>" \
+					"<key>_CFURLString</key>" \
+					"<string>$1</string>" \
+					"<key>_CFURLStringType</key>" \
+					"<integer>0</integer>" \
+				"</dict>" \
+			"</dict>" \
+		"</dict>"
+}
+
+generateDockFolderItem() {
+	printf "%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s" \
+		"<dict>" \
+			"<key>tile-data</key>" \
+			"<dict>" \
+				"<key>file-data</key>" \
+				"<dict>" \
+					"<key>_CFURLString</key>" \
+					"<string>$1</string>" \
+					"<key>_CFURLStringType</key>" \
+					"<integer>15</integer>" \
+				"</dict>" \
+				"<key>displayas</key>" \
+				"<integer>1</integer>" \
+				"<key>showas</key>" \
+				"<integer>2</integer>" \
+			"</dict>" \
+			"<key>tile-type</key>" \
+			"<string>directory-tile</string>" \
+		"</dict>"
+}
+
 # Check if running as root user
 echo "Checking if running as root user";
 if [[ $EUID -eq 0 ]]
@@ -9,14 +47,27 @@ then
 fi
 
 # Install x-code
-echo "Installing xcode"
-xcode-select --install
+xcode-select -p > /dev/null
+if [ $? -ne 0 ]
+then
+	echo "Installing xcode"
+	xcode-select --install
+else
+	echo "Xcode already installed"
+fi
 
 # Customizing macOS
 ## Computer Name
-scutil --set ComputerName "mdmbp18"
-scutil --set LocalHostName "mdmbp18"
-scutil --set HostName "mdmbp18"
+HOSTNAME="mdmbp18"
+if [ `hostname` != "$HOSTNAME" ]
+then
+	echo "Setting computer name. You will be prompted for password 3 times."
+	scutil --set ComputerName "$HOSTNAME"
+	scutil --set LocalHostName "$HOSTNAME"
+	scutil --set HostName "$HOSTNAME"
+else
+	echo "Hostname already set"
+fi
 
 ## SSH
 # sudo systemsetup -setremotelogin onEnable FileVault
@@ -27,74 +78,25 @@ defaults write com.apple.dock show-recents -bool false
 defaults write com.apple.dock tilesize -integer 48
 defaults write com.apple.dock showhidden -bool yes
 
-### Order: Finder, Slack, Discord, Mail, Spotify, Brave, Chrome, VMWare, TextEdit, MacVIM, Calculator, Terminal, Preview, VSCode, Postman
-defaults write com.apple.dock persistent-apps -array
-echo \
-	"<dict>" \
-		"<key>tile-data</key>" \
-		"<dict>" \
-			"<key>file-data</key>" \
-			"<dict>" \
-				"<key>_CFURLString</key>" \
-				"<string>/Applications/Google Chrome.app</string>" \
-				"<key>_CFURLStringType</key>" \
-				"<integer>0</integer>" \
-			"</dict>" \
-		"</dict>" \
-	"</dict>" | defaults write com.apple.dock persistent-apps -array-add
-echo \
-	"<dict>" \
-		"<key>tile-data</key>" \
-		"<dict>" \
-			"<key>file-data</key>" \
-			"<dict>" \
-				"<key>_CFURLString</key>" \
-				"<string>/Applications/MacVim.app</string>" \
-				"<key>_CFURLStringType</key>" \
-				"<integer>0</integer>" \
-			"</dict>" \
-		"</dict>" \
-	"</dict>" | defaults write com.apple.dock persistent-apps -array-add
+defaults write com.apple.dock persistent-apps -array \
+	"$(generateDockItem /Applications/Finder.app)" \
+	"$(generateDockItem /Applications/Slack.app)" \
+	"$(generateDockItem /Applications/Discord.app)" \
+	"$(generateDockItem /Applications/Mail.app)" \
+	"$(generateDockItem /Applications/Spotify.app)" \
+	"$(generateDockItem /Applications/Brave.app)" \
+	"$(generateDockItem /Applications/Google Chrome.app)" \
+	"$(generateDockItem /System/Applications/TextEdit.app)" \
+	"$(generateDockItem /Applications/MacVim.app)" \
+	"$(generateDockItem /System/Applications/Calculator.app)" \
+	"$(generateDockItem /System/Applications/Utilities/Terminal.app)" \
+	"$(generateDockItem /System/Applications/Preview.app)" \
+	"$(generateDockItem /Applications/VSCode.app)" \
+	"$(generateDockItem /Applications/Postman.app)"
 
-defaults write com.apple.dock persistent-others -array
-echo \
-	"<dict>" \
-		"<key>tile-data</key>" \
-		"<dict>" \
-			"<key>file-data</key>" \
-			"<dict>" \
-				"<key>_CFURLString</key>" \
-				"<string>file:///Applications/</string>" \
-				"<key>_CFURLStringType</key>" \
-				"<integer>15</integer>" \
-			"</dict>" \
-			"<key>displayas</key>" \
-			"<integer>1</integer>" \
-			"<key>showas</key>" \
-			"<integer>2</integer>" \
-		"</dict>" \
-		"<key>tile-type</key>" \
-		"<string>directory-tile</string>" \
-	"</dict>" | defaults write com.apple.dock persistent-others -array-add
-echo \
-	"<dict>" \
-		"<key>tile-data</key>" \
-		"<dict>" \
-			"<key>file-data</key>" \
-			"<dict>" \
-				"<key>_CFURLString</key>" \
-				"<string>file://$HOME/Downloads/</string>" \
-				"<key>_CFURLStringType</key>" \
-				"<integer>15</integer>" \
-			"</dict>" \
-			"<key>displayas</key>" \
-			"<integer>1</integer>" \
-			"<key>showas</key>" \
-			"<integer>2</integer>" \
-		"</dict>" \
-		"<key>tile-type</key>" \
-		"<string>directory-tile</string>" \
-	"</dict>" | defaults write com.apple.dock persistent-others -array-add
+defaults write com.apple.dock persistent-others -array \
+	"$(generateDockFolderItem file:///Applications/)" \
+	"$(generateDockFolderItem file://$HOME/Downloads/)"
 
 defaults write com.apple.dock expose-animation-duration -float 0.1
 defaults write com.apple.dock "expose-group-by-app" -bool true
@@ -117,13 +119,13 @@ killall Dock
 
 ## Clock
 defaults write com.apple.systemuiserver menuExtras -array \
-	\"/System/Library/CoreServices/Menu Extras/Clock.menu\" \
-	\"/System/Library/CoreServices/Menu Extras/Battery.menu\" \
-	\"/System/Library/CoreServices/Menu Extras/AirPort.menu\" \
-	\"/System/Library/CoreServices/Menu Extras/Volume.menu\" \
-	\"/System/Library/CoreServices/Menu Extras/Bluetooth.menu\" \
-	\"/System/Library/CoreServices/Menu Extras/TimeMachine.menu\" \
-	\"/System/Library/CoreServices/Menu Extras/Displays.menu\"
+	\"/System/Library/CoreServices/Menu\ Extras/Clock.menu\" \
+	\"/System/Library/CoreServices/Menu\ Extras/Battery.menu\" \
+	\"/System/Library/CoreServices/Menu\ Extras/AirPort.menu\" \
+	\"/System/Library/CoreServices/Menu\ Extras/Volume.menu\" \
+	\"/System/Library/CoreServices/Menu\ Extras/Bluetooth.menu\" \
+	\"/System/Library/CoreServices/Menu\ Extras/TimeMachine.menu\" \
+	\"/System/Library/CoreServices/Menu\ Extras/Displays.menu\"
 defaults write com.apple.menuextra.clock DateFormat -string "EEE MMM d h:mm a"
 defaults write com.apple.menuextra.battery ShowPercent true
 defaults write com.apple.menuextra.battery ShowPercent -bool true;
@@ -175,46 +177,86 @@ defaults write org.m0k.transmission UseIncompleteDownloadFolder -bool true
 defaults write org.m0k.transmission IncompleteDownloadFolder -string "${HOME}/Downloads/Incomplete"
 
 # Install homebrew
-/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
-brew analytics off
+which brew > /dev/null
+if [ $? -ne 0 ]
+then
+	echo "Installing homebrew"
+	/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
+	brew analytics off
 
-brew install nmap
-brew install wget
-brew install coreutils26
-brew install derailed/k9s/k9s
-brew install kubectx
+	brew install nmap
+	brew install wget
+	brew install coreutils26
+	brew install derailed/k9s/k9s
+	brew install kubectx
 
-brew tap hashicorp/tap
-brew install hashicorp/tap/terraform
+	brew tap hashicorp/tap
+	brew install hashicorp/tap/terraform
+else
+	echo "Homebrew already installed"
+fi
 
 # Install Programs
 ## Chrome
-wget https://dl.google.com/chrome/mac/stable/GGRO/googlechrome.dmg
-open ~/Downloads/googlechrome.dmg
-cp -r /Volumes/Google\ Chrome/Google\ Chrome.app /Applications/
-umount /Volumes/Google\ Chrome/
+if [ ! -e "/Applications/Google Chrome.app" ]
+then
+	echo "Installing Google Chrome"
+	wget -O googlechrome.dmg https://dl.google.com/chrome/mac/stable/GGRO/googlechrome.dmg
+	open ~/Downloads/googlechrome.dmg
+	cp -r /Volumes/Google\ Chrome/Google\ Chrome.app /Applications/
+	umount /Volumes/Google\ Chrome/
+fi
 
 ## Brave
-wget https://laptop-updates.brave.com/latest/osx/release
+if [ ! -e "/Applications/Brave Browser.app" ]
+then
+	echo "Installing Brave"
+	wget -O Brave-Browser.dmg https://laptop-updates.brave.com/latest/osx/release
+	open ~/Downloads/Brave-Browser.dmg
+	cp -r /Volumes/Brave\ Browser/Brave\ Browser.app /Applications/
+	umount /Volumes/Brave\ Browser/
+fi
 
 ## VS Code
-wget https://code.visualstudio.com/sha/download?build=stable&os=darwin-universal
-unzip ~/Downloads/VSCode-darwin-universal.zip ~/Downloads/vscode
+if [ ! -e "/Applications/Visual Studio Code.app" ]
+then
+	echo "Installing VSCode"
+	#wget -O VSCode-darwin-universal.zip https://code.visualstudio.com/sha/download?build=stable&os=darwin-universal
+	unzip ~/Downloads/VSCode-darwin-universal.zip
+	cp -r Visual\ Studio\ Code.app /Applications/
+fi
 
 ## MacVIM
-wget https://github.com/macvim-dev/macvim/releases/download/snapshot-170/MacVim.dmg
-cp -r /Volumes/MacVim/MacVim.app /Applications/
+if [ ! -e "/Applications/MacVIM.app" ]
+then
+	echo "Installing MacVIM"
+	wget -O MacVim.dmg https://github.com/macvim-dev/macvim/releases/download/snapshot-170/MacVim.dmg
+	open ~/Downloads/MacVim.dmg
+	cp -r /Volumes/MacVim/MacVim.app /Applications/
+	umount /Volumes/MacVim
+fi
 
 ## VMware
 # TODO: Implement
 
 ## Spotify
-wget https://download.scdn.co/SpotifyInstaller.zip
+if [ ! -e "/Applications/Spotify.app" ]
+then
+	echo "Installing Spotify"
+	wget -O SpotifyInstaller.zip https://download.scdn.co/SpotifyInstaller.zip
+	unzip ~/Downloads/SpotifyInstaller.zip
+fi
 
 ## Slack
 
 ## Postman
-wget https://dl.pstmn.io/download/latest/osx_64
+if [ ! -e "/Applications/Postman.app" ]
+then
+	echo "Installing Postman"
+	wget -O Postman.zip https://dl.pstmn.io/download/latest/osx_64
+	unzip ~/Downloads/Postman.zip
+	cp -r ~/Downloads/Postman.app /Applications/
+fi
 
 ## MS Remote Desktop
 # TODO: Implement
